change: clean up configuration code
