+++ server/configuration/loader.go
@@ -1,8 +1,6 @@
 package configuration
 
 import (
-	"errors"
-	"fmt"
 	"log/slog"
 	"os"
 	"strings"
@@ -14,38 +12,7 @@ import (
 	"github.com/knadh/koanf/v2"
 )
 
-var (
-	systemConfiguration Configuration
-	k                   = koanf.New(".")
-	validSystemModes    = map[string]bool{
-		"DEV":  true,
-		"PROD": true,
-	}
-	validSystemLogLevels = map[string]bool{
-		"INFO":  true,
-		"WARN":  true,
-		"ERROR": true,
-	}
-	validLanguages = map[string]bool{
-		"en": true,
-		"de": true,
-	}
-)
-
-func Get() Configuration {
-	return systemConfiguration
-}
-
-func Init() {
-	loadConfigurationFromJson()
-	loadConfigurationFromEnv()
-	loadMissingConfigurationValues()
-
-	loadIntoGlobalStruct()
-	validateConfiguration()
-
-	applyLogLevel()
-}
+var k = koanf.New(".")
 
 func loadConfigurationFromJson() {
 	localConfigPath := "../config.json"
@@ -190,44 +157,6 @@ func loadIntoGlobalStruct() {
 	}
 }
 
-func validateConfiguration() {
-	var configErrors []error
-
-	if systemConfiguration.Database.Password == "" {
-		configErrors = append(configErrors, fmt.Errorf("database.password is empty, please set a password"))
-	}
-
-	if systemConfiguration.System.Domain == "" {
-		configErrors = append(configErrors, fmt.Errorf("system.domain is required"))
-	}
-
-	if !validSystemModes[systemConfiguration.System.Mode] {
-		configErrors = append(configErrors, errors.New("system.mode has to be either 'DEV' or 'PROD'"))
-	}
-
-	if systemConfiguration.System.LogLevel != "" && !validSystemLogLevels[systemConfiguration.System.LogLevel] {
-		configErrors = append(configErrors, errors.New("system.loglevel has to be either unset, 'INFO', 'WARN' or 'ERROR"))
-	}
-
-	if systemConfiguration.Admin.Password == "" {
-		configErrors = append(configErrors, fmt.Errorf("admin.password is empty, please set a password"))
-	}
-
-	if !validLanguages[systemConfiguration.System.Frontend.DefaultLanguage] {
-		configErrors = append(configErrors, fmt.Errorf("system.frontend.default_language has to be either 'en', or 'de'"))
-	}
-
-	if len(configErrors) > 0 {
-		slog.Error("the configration has several errors:")
-		for _, err := range configErrors {
-			slog.Error(err.Error())
-		}
-
-		slog.Error("Software booting aborted due to configuration errors")
-		os.Exit(1)
-	}
-}
-
 func applyLogLevel() {
 	switch systemConfiguration.System.LogLevel {
 	case "INFO":
@@ -237,43 +166,4 @@ func applyLogLevel() {
 	case "ERROR":
 		utils.LogLevel.Set(slog.LevelError)
 	}
-}
-
-type Configuration struct {
-	Admin struct {
-		Mail     string `koanf:"mail"`
-		Password string `koanf:"password"`
-	} `koanf:"admin"`
-	Database struct {
-		Host     string `koanf:"host"`
-		Port     string `koanf:"port"`
-		User     string `koanf:"user"`
-		Password string `koanf:"password"`
-		Name     string `koanf:"db"`
-	} `koanf:"database"`
-	System struct {
-		Domain   string `koanf:"domain"`
-		Mode     string `koanf:"mode"`
-		Pepper   string `koanf:"pepper"`
-		LogLevel string `koanf:"loglevel"`
-		Frontend struct {
-			DefaultLanguage string `koanf:"default_language"`
-			MaxInputs       struct {
-				Public struct {
-					Title   int `koanf:"title"`
-					Content int `koanf:"content"`
-				} `koanf:"public"`
-				Private struct {
-					Titles int `koanf:"titles"`
-					Labels int `koanf:"labels"`
-					Names  int `koanf:"names"`
-					Faqs   struct {
-						Questions int `koanf:"questions"`
-						Answers   int `koanf:"answers"`
-					} `koanf:"faqs"`
-					About int `koanf:"about"`
-				} `koanf:"private"`
-			} `koanf:"max_inputs"`
-		} `koanf:"frontend"`
-	} `koanf:"system"`
 }
